refactor(delegate_to): delete presenter
