+++ lib/convenient_service/rspec/matchers/custom/delegate_to/entities/matcher/commands.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-require_relative "commands/apply_stub_to_track_delegations"
+require_relative "commands/generate_printable_method"
