+++ lib/convenient_service/rspec/matchers/custom/delegate_to/entities/matcher/entities.rb
@@ -3,4 +3,3 @@
 require_relative "entities/chainings"
 require_relative "entities/chainings_collection"
 require_relative "entities/delegation"
-require_relative "entities/presenter"
