--- lib/convenient_service/rspec/matchers/custom/delegate_to/entities/matcher/entities/presenter/commands.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "commands/generate_printable_method"
