+++ spec/lib/convenient_service/rspec/matchers/custom/delegate_to/entities/matcher/commands/generate_printable_method_spec.rb
@@ -5,7 +5,7 @@ require "spec_helper"
 require "convenient_service"
 
 # rubocop:disable RSpec/NestedGroups, RSpec/MultipleMemoizedHelpers
-RSpec.describe ConvenientService::RSpec::Matchers::Custom::DelegateTo::Entities::Matcher::Entities::Presenter::Commands::GeneratePrintableMethod do
+RSpec.describe ConvenientService::RSpec::Matchers::Custom::DelegateTo::Entities::Matcher::Commands::GeneratePrintableMethod do
   example_group "class methods" do
     describe ".call" do
       let(:command_result) { described_class.call(object: object, method: method) }
